requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.7.0

# LLM + Orchestration
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0

# Graph DB
neo4j>=5.20.0

# Vector DB
faiss-cpu>=1.8.0
numpy>=1.26.0

# Document parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0

# Tokenization
tiktoken>=0.7.0

# Optional: tracing
langsmith>=0.1.0
